Reorganize functions and traits.
Group related things and move ToLowerCharVec to lib.rs.
